encapsulated manipulation of problem's data
